fixed a logical statement for removing dead links for udev systemd files in 61_verify_and_adjust_udev_systemd.sh layout/prepare/GNU/Linux/13_include_filesystem_code.sh added a forgotten mount statement for btrfs without subvolumes

+++ usr/share/rear/build/GNU/Linux/61_verify_and_adjust_udev_systemd.sh
@@ -16,7 +16,7 @@ Log "Cleaning up systemd udev socket files"
 my_udev_files=( $(find $ROOTFS_DIR/usr/lib/systemd/system/sockets.target.wants -type l -name "*udev*"  -printf "%P\n") )
 
 for m in "${my_udev_files[@]}" ; do
-    if [[ ! -h /lib/systemd/system/sockets.target.wants/$m ]] || [[ ! -h /usr/lib/systemd/system/sockets.target.wants/$m ]]; then
+    if [[ ! -h /lib/systemd/system/sockets.target.wants/$m ]] && [[ ! -h /usr/lib/systemd/system/sockets.target.wants/$m ]]; then
         rm -f $ROOTFS_DIR/usr/lib/systemd/system/sockets.target.wants/$m
     fi
 done
